+++ lib/skooma/objects/parameter/keywords/content.rb
@@ -8,6 +8,7 @@ module Skooma
           self.key = "content"
           self.value_schema = :object_of_schemas
           self.schema_value_class = Objects::MediaType
+          self.depends_on = %w[in name style explode allowReserved allowEmptyValue]
 
           def evaluate(instance, result)
             return if instance.value.nil?