+++ lib/skooma/objects/parameter/keywords/required.rb
@@ -6,6 +6,7 @@ module Skooma
       module Keywords
         class Required < JSONSkooma::Keywords::Base
           self.key = "required"
+          self.depends_on = %w[in name style explode allowReserved allowEmptyValue]
 
           def evaluate(instance, result)
             if json.value && ValueParser.call(instance, result)&.value.nil?