+++ examples/test_app.rb
@@ -3,21 +3,25 @@
 require "sinatra"
 require "json"
 
-TestApp = Sinatra.new do
-  get "/" do
-    content_type :json
-    JSON.generate({"foo" => "bar"})
-  end
+module TestApp
+  def self.[](value)
+    Sinatra.new do
+      get "/" do
+        content_type :json
+        JSON.generate({"foo" => value})
+      end
 
-  post "/" do
-    content_type :json
-    data = JSON.parse request.body.read
-    if data["foo"] == "bar"
-      status 201
-      JSON.generate({"foo" => "bar"})
-    else
-      status 400
-      JSON.generate({"message" => "foo must be bar"})
+      post "/" do
+        content_type :json
+        data = JSON.parse request.body.read
+        if data["foo"] == value
+          status 201
+          JSON.generate({"foo" => value})
+        else
+          status 400
+          JSON.generate({"message" => "foo must be #{value}"})
+        end
+      end
     end
   end
 end