+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning].
 
 ## [Unreleased]
 
+## [0.2.3] - 2024-01-18
+
 ### Added
 
 - Add support for multiple OpenAPI documents. ([@skryukov])
@@ -67,7 +69,8 @@ end
 
 [@skryukov]: https://github.com/skryukov
 
-[Unreleased]: https://github.com/skryukov/skooma/compare/v0.2.2...HEAD
+[Unreleased]: https://github.com/skryukov/skooma/compare/v0.2.3...HEAD
+[0.2.3]: https://github.com/skryukov/skooma/compare/v0.2.2...v0.2.3
 [0.2.2]: https://github.com/skryukov/skooma/compare/v0.2.1...v0.2.2
 [0.2.1]: https://github.com/skryukov/skooma/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/skryukov/skooma/compare/v0.1.0...v0.2.0